requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
youtube-transcript-api==0.6.2
openai-whisper==20231117
yt-dlp==2023.11.16
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
